Update behaviour to match test
- Hi David, I hope you don't mind me butting in again but I had an
'unfinished' feeling nagging at the back of my mind
so I came back to see if I could get the test working
- the ConversationsPage was not showing the displayNames anywhere so
I changed the ConversationsListTile to show the combined set of display
names as the title
- i changed the 'expect' call to look for the combined set of display
names
- I imagine we will change the ConversationsListTile to use a different
title at some point, eg. a truncated version of the combined list of
display names. I suggest we add a 'getTruncatedNames()' function to the
DatabaseService and move the viewmodel state inside the DatabaseService,
so we can use the function in the test but i imagine that would be a
different PR
- thanks for working with me on this!
-